Fix static tests on pull request #1866 (https://github.com/magento-engcom/msi/pull/1866)
